test unitaire done

+++ src/main/kotlin/org/ldv/sushi/apisushi/dto/AlimentBoxDtoJson.kt
@@ -2,7 +2,7 @@ package org.ldv.sushi.apisushi.dto
 
 class AlimentBoxDtoJson (
     var nom: String,
-    var quantite: Int) {
+    var quantite: Double) {
 
     override fun toString(): String {
         return "AlimentBoxDtoJson(nom='$nom', quantite=$quantite)"
